requirements.txt
bootstrap4==0.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dash==1.1.1
dash-core-components==1.1.1
dash-html-components==1.0.0
dash-renderer==1.0.0
dash-table==4.1.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.1
django-allauth==0.39.1
django-bootstrap==0.2.4
django-bootstrap-calendar==0.1.1
django-bootstrap-datepicker==1.3.1
django-bootstrap-pagination==1.7.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-hitcount==1.3.1
django-multiselectfield==0.1.8
django-mysql==3.2.0
django-qrcode==0.3
djangorestframework==3.10.2
Flask==1.1.1
Flask-Compress==1.4.0
google-api-python-client==1.4.1
gunicorn==19.9.0
httplib2==0.13.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jsmin==2.2.2
lxml==4.4.1
MarkupSafe==1.1.1
mtnmomo==3.0.1
numpy==1.17.0
oauth2client==4.1.3
oauthlib==3.0.1
pandas==0.25.0
pandas-datareader==0.7.4
phonenumbers==8.10.10
Pillow==6.0.0
plotly==4.1.0
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.2
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
rsa==4.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
utils==0.9.0
Werkzeug==0.15.5
whitenoise==4.1.3
wrapt==1.11.2
